+++ src/app/clients/edit-client/edit-client.component.ts
@@ -7,6 +7,10 @@ import { UntypedFormBuilder, UntypedFormGroup, Validators, UntypedFormControl } 
 import { ClientsService } from '../clients.service';
 import { SettingsService } from 'app/settings/settings.service';
 import { Dates } from 'app/core/utils/dates';
+import { ClientOtpDialogComponent } from '../client-otp-dialog/client-otp-dialog.component';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { SystemService } from 'app/system/system.service';
+import { AlertService } from 'app/core/alert/alert.service';
 import { MatomoTracker } from "@ngx-matomo/tracker";
 
 /**
@@ -62,8 +66,10 @@ export class EditClientComponent implements OnInit {
     private clientsService: ClientsService,
     private dateUtils: Dates,
     private settingsService: SettingsService,
-    private matomoTracker: MatomoTracker
-  ) {
+    private matomoTracker: MatomoTracker,
+    private dialog: MatDialog,
+    private systemService: SystemService,
+    private alertService: AlertService) {
     this.route.data.subscribe((data: { clientDataAndTemplate: any }) => {
       this.clientDataAndTemplate = data.clientDataAndTemplate;
     });
@@ -164,7 +170,7 @@ export class EditClientComponent implements OnInit {
   /**
    * Submits the edit client form.
    */
-  submit() {
+  private updateClient() {
     const locale = this.settingsService.language.code;
     const dateFormat = this.settingsService.dateFormat;
     const editClientFormValue: any = this.editClientForm.getRawValue();
@@ -200,4 +206,85 @@ export class EditClientComponent implements OnInit {
     });
   }
 
+  /**
+   * Opens OTP dialog to validate phone number.
+   * @param phoneNumber The client's phone number that recieved the OTP
+   * @param countryId The client's country ID
+   */
+  private openOtpDialog(phoneNumber: string, countryId: number): void {
+    this.systemService.getConfigurationByName('country-client-phone-number-otp-length', { countryId })
+    .subscribe(config => {
+      if (config?.enabled && config?.value > 3) {
+        const otpDialog = this.dialog.open(ClientOtpDialogComponent, {
+          data: { mobileNo: phoneNumber,
+                  countryId: countryId,
+                  otpLength: config.value
+           }
+        });
+        this.updateAfterOtpValidation(otpDialog);
+      } else {
+        this.alertService.alert({
+          type: 'Configuration Error',
+          message: 'Invalid OTP length configuration. Please contact your administrator.'
+        });
+      }
+    });
+  }
+
+  /**
+   *
+   * Submits the request to update client after OTP validation. The OTP dialog is opened only if the phone number is updated and OTP validation is required in the client's country.
+   */
+  submit(): void {
+    let formMobileNo: string = this.editClientForm.get('mobileNo').value;
+    if (formMobileNo) {
+      formMobileNo = formMobileNo.trim();
+    }
+    if (formMobileNo === this.clientDataAndTemplate.mobileNo) {
+      this.updateClient();
+      return;
+    }
+    const countryId = this.clientDataAndTemplate.countryId;
+    this.systemService.getConfigurationByName('country-client-identity-ocr-validation-required', { countryId })
+    .subscribe(config => {
+      if (config?.enabled) {
+        this.initiateOtpValidation(formMobileNo, countryId);
+      } else {
+        this.updateClient();
+      }
+    })
+  }
+
+  /**
+   * Updates client after OTP validation.
+   * @param otpDialog OTP Dialog that got opened
+   */
+  private updateAfterOtpValidation(otpDialog: MatDialogRef<ClientOtpDialogComponent, any>): void {
+    otpDialog.afterClosed().subscribe(validOtp => {
+      if (validOtp) {
+        this.updateClient();
+      }
+    });
+  }
+
+  /**
+   * Initiates OTP validation.
+   * @param mobileNo The client's phone number
+   * @param countryId The client's country ID
+   */
+  private initiateOtpValidation(mobileNo: string, countryId: number): void {
+    if (!mobileNo) {
+      this.alertService.alert({
+        type: 'Validation Error',
+        message: 'Please provide a valid phone number for OTP validation.'
+      });
+      return;
+    }
+    if (!mobileNo.startsWith('+')) {
+      mobileNo = `+${mobileNo}`;
+    }
+    this.clientsService.generateClientOTP(countryId, {mobilePhoneNumber: mobileNo}).subscribe(() => {
+      this.openOtpDialog(mobileNo, countryId);
+    });
+  }
 }